Create Complete files
The first file contain all the code needed to create and train any of the models. The second file contains all the code needed to obtain the visualization tools. Those files are ready to be ran (provided that the paths are adapted to fit the cluster on which the code is run).